feat: move order card to package components
